change password success

+++ crowdfunding/echo/views.py
@@ -1,4 +1,5 @@
 from rest_framework.views import APIView
+from rest_framework import generics
 from rest_framework import status, permissions
 from django.http import Http404
 from rest_framework.response import Response
@@ -8,7 +9,7 @@ from .serializers import EchoSerializer, PledgeSerializer,EchoDetailSerializer
 from .permissions import IsOwnerOrReadOnly
 
 #configure to view all projects
-class ProjectList(APIView):
+class ProjectList(APIView): #APIView
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def get(self,request):
@@ -24,12 +25,20 @@ class ProjectList(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 #configure to view specific project
-class ProjectDetail(APIView):
+class ProjectDetail(generics.RetrieveUpdateDestroyAPIView): #APIView
     permission_classes = [permissions.IsAuthenticatedOrReadOnly, IsOwnerOrReadOnly]
 
-    def get_object(self,pk):
+    # def get_object(self,pk):
+    #     try:
+    #         return Echo.objects.get(pk=pk)
+    #     except Echo.DoesNotExist:
+    #         raise Http404
+
+    def get_object(self, pk):
         try:
-            return Echo.objects.get(pk=pk)
+            project = Echo.objects.get(pk=pk)
+            self.check_object_permissions(self.request, project)
+            return project
         except Echo.DoesNotExist:
             raise Http404
 
@@ -52,9 +61,13 @@ class ProjectDetail(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request,pk):
-        project = self.get_object(pk)
+        project = self.objects.get(pk=pk)
+        self.check_object_permissions(self.request, project)
+        #project = self.get_object(pk)
+        #how to add condition so that only owner can delete its own project, and if there are no supporters
         project.delete()
         return Response(status=status.HTTP_200_OK)
+        #return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 #configure to view all pledges
 class PledgeList(APIView):
@@ -73,10 +86,17 @@ class PledgeList(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 #configure to view specific pledge
-class PledgeDetail(APIView):
-    def get_object(self,pk):
+class PledgeDetail(generics.RetrieveAPIView):
+    # def get_object(self,pk):
+    #     try:
+    #         return Pledge.objects.get(pk=pk)
+    #     except Pledge.DoesNotExist:
+    #         raise Http404
+    def get_object(self, pk):
         try:
-            return Pledge.objects.get(pk=pk)
+            pledge = Pledge.objects.get(pk=pk)
+            self.check_object_permissions(self.request, pledge)
+            return pledge
         except Pledge.DoesNotExist:
             raise Http404
